backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.35
psycopg[binary]>=3.1.18
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
aiofiles>=23.2.1
pandas>=2.2.0
numpy>=1.26.4
scipy>=1.11.4
statsmodels>=0.14.0
scikit-learn>=1.3.2
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
joblib>=1.3.0
python-multipart>=0.0.6
prophet>=1.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
groq>=0.11.0
reportlab>=4.0.0
python-pptx>=0.6.23
python-docx>=1.1.0
pymongo>=4.6.0
pymysql>=1.1.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
apscheduler>=3.10.4
duckdb>=0.9.2
openpyxl>=3.1.2
xlrd>=2.0.1
chardet>=5.2.0
